# src/shared/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ALAssLine": {
				"properties": {
					"ALAssLineID": {
						"type": "number",
						"format": "double"
					},
					"ProductID": {
						"type": "number",
						"format": "double"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					},
					"Name": {
						"type": "string"
					},
					"Status": {
						"$ref": "#/components/schemas/ALAssLineStatus"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"allocations": {
						"items": {
							"$ref": "#/components/schemas/ALAssLineWStationAllocation"
						},
						"type": "array"
					}
				},
				"required": [
					"ALAssLineID",
					"ProductID",
					"product",
					"Name",
					"Status",
					"createdAt",
					"updatedAt",
					"allocations"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Product": {
				"properties": {
					"_executorEmail": {
						"type": "string"
					},
					"ProductID": {
						"type": "number",
						"format": "double"
					},
					"Name": {
						"type": "string"
					},
					"Active": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"assemblyLines": {
						"items": {
							"$ref": "#/components/schemas/ALAssLine"
						},
						"type": "array"
					}
				},
				"required": [
					"ProductID",
					"Name",
					"Active",
					"createdAt",
					"updatedAt",
					"assemblyLines"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ALAssLineStatus": {
				"enum": [
					1,
					2,
					3
				],
				"type": "number"
			},
			"ALAssLineWStationAllocation": {
				"properties": {
					"ALAssLineWStationAllocationID": {
						"type": "number",
						"format": "double"
					},
					"ALAssLineID": {
						"type": "number",
						"format": "double"
					},
					"ALWStationID": {
						"type": "number",
						"format": "double"
					},
					"Sort": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"assemblyLine": {
						"$ref": "#/components/schemas/ALAssLine"
					},
					"workstation": {
						"$ref": "#/components/schemas/ALWStation"
					}
				},
				"required": [
					"ALAssLineWStationAllocationID",
					"ALAssLineID",
					"ALWStationID",
					"Sort",
					"createdAt",
					"updatedAt",
					"assemblyLine",
					"workstation"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ALWStation": {
				"properties": {
					"ALWStationID": {
						"type": "number",
						"format": "double"
					},
					"Name": {
						"type": "string"
					},
					"ShortName": {
						"type": "string"
					},
					"PCName": {
						"type": "string"
					},
					"AutoStart": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"allocations": {
						"items": {
							"$ref": "#/components/schemas/ALAssLineWStationAllocation"
						},
						"type": "array"
					}
				},
				"required": [
					"ALWStationID",
					"Name",
					"ShortName",
					"PCName",
					"AutoStart",
					"createdAt",
					"updatedAt",
					"allocations"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiError": {
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiMeta": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"links": {
						"properties": {
							"details": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"entity": {
						"properties": {
							"primaryKey": {
								"type": "string"
							}
						},
						"required": [
							"primaryKey"
						],
						"type": "object"
					}
				},
				"required": [
					"page",
					"limit",
					"total"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseList_ALWStation_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ALWStation"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					},
					"meta": {
						"$ref": "#/components/schemas/ApiMeta"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseSingle_ALWStation_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ALWStation"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseInfo": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponse_ALWStation_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ApiResponseSingle_ALWStation_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseList_ALWStation_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseInfo"
					},
					{
						"$ref": "#/components/schemas/ApiError"
					}
				]
			},
			"GetProductsSchema": {
				"properties": {
					"ProductID": {
						"type": "number",
						"format": "double"
					},
					"Name": {
						"type": "string"
					},
					"Active": {
						"type": "number",
						"format": "double"
					},
					"CreatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"UpdatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"ProductID",
					"Name",
					"Active",
					"CreatedAt",
					"UpdatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseSingle_GetProductsSchema_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/GetProductsSchema"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseList_GetProductsSchema_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/GetProductsSchema"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					},
					"meta": {
						"$ref": "#/components/schemas/ApiMeta"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponse_GetProductsSchema_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ApiResponseSingle_GetProductsSchema_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseList_GetProductsSchema_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseInfo"
					},
					{
						"$ref": "#/components/schemas/ApiError"
					}
				]
			},
			"GetProductSchema": {
				"properties": {
					"ProductID": {
						"type": "number",
						"format": "double"
					},
					"Name": {
						"type": "string"
					},
					"Active": {
						"type": "number",
						"format": "double"
					},
					"CreatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"UpdatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"ProductID",
					"Name",
					"Active",
					"CreatedAt",
					"UpdatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseSingle_GetProductSchema_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/GetProductSchema"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseSingle_Product_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/Product"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseList_Product_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					},
					"meta": {
						"$ref": "#/components/schemas/ApiMeta"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponse_Product_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ApiResponseSingle_Product_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseList_Product_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseInfo"
					},
					{
						"$ref": "#/components/schemas/ApiError"
					}
				]
			},
			"PostProductsSchema": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Active": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"Name",
					"Active"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PatchProductsSchema": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Active": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"Name",
					"Active"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseList_ALAssLine_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ALAssLine"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					},
					"meta": {
						"$ref": "#/components/schemas/ApiMeta"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseSingle_ALAssLine_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ALAssLine"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponse_ALAssLine_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ApiResponseSingle_ALAssLine_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseList_ALAssLine_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseInfo"
					},
					{
						"$ref": "#/components/schemas/ApiError"
					}
				]
			},
...
			"ApiResponseList_ALAssLineWStationAllocation_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ALAssLineWStationAllocation"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					},
					"meta": {
						"$ref": "#/components/schemas/ApiMeta"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponseSingle_ALAssLineWStationAllocation_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ALAssLineWStationAllocation"
					},
					"error": {
						"$ref": "#/components/schemas/ApiError"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ApiResponse_ALAssLineWStationAllocation_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ApiResponseSingle_ALAssLineWStationAllocation_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseList_ALAssLineWStationAllocation_"
					},
					{
						"$ref": "#/components/schemas/ApiResponseInfo"
					},
					{
						"$ref": "#/components/schemas/ApiError"
					}
				]
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
		"/api/workstations": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseList_ALWStation_"
								}
							}
						}
					}
				},
				"tags": [
					"Workstations"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ALWStation_"
								}
							}
						}
					}
				},
				"tags": [
					"Workstations"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"PCName": {
										"type": "string"
									},
									"ShortName": {
										"type": "string"
									},
									"Name": {
										"type": "string"
									}
								},
								"required": [
									"PCName",
									"ShortName",
									"Name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/workstations/{id}": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ALWStation_"
								}
							}
						}
					}
				},
				"tags": [
					"Workstations"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"PCName": {
										"type": "string"
									},
									"ShortName": {
										"type": "string"
									},
									"Name": {
										"type": "string"
									}
								},
								"required": [
									"PCName",
									"ShortName",
									"Name"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Workstations"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/products": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetProductsSchema_"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Product_"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostProductsSchema"
							}
						}
					}
				}
			}
		},
		"/api/products/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseSingle_GetProductSchema_"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Product_"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PatchProductsSchema"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/lines": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseList_ALAssLine_"
								}
							}
						}
					}
				},
				"tags": [
					"Lines"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ALAssLine_"
								}
							}
						}
					}
				},
				"tags": [
					"Lines"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ALAssLine"
							}
						}
					}
				}
			}
		},
		"/api/lines/{id}": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ALAssLine_"
								}
							}
						}
					}
				},
				"tags": [
					"Lines"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ALAssLine"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Lines"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/allocations": {
			"get": {
				"operationId": "GetAllocations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseList_ALAssLineWStationAllocation_"
								}
							}
						}
					}
				},
				"tags": [
					"Allocations"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateAllocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ALAssLineWStationAllocation_"
								}
							}
						}
					}
				},
				"tags": [
					"Allocations"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"ALWStationID": {
										"type": "number",
										"format": "double"
									},
									"ALAssLineID": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"ALWStationID",
									"ALAssLineID"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Allocations"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ids",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"type": "number",
								"format": "double"
							}
						}
					}
				]
			}
		},
		"/api/allocations/{id}": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ALAssLineWStationAllocation_"
								}
							}
						}
					}
				},
				"tags": [
					"Allocations"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"Sort": {
										"type": "number",
										"format": "double"
									},
									"ALWStationID": {
										"type": "number",
										"format": "double"
									},
									"ALAssLineID": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"Sort",
									"ALWStationID",
									"ALAssLineID"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Allocations"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "https://[object Object]/"
		}
	]
}